Feature/modernize parser class

+++ hasm/include/hasm/ErrorMessage.hpp
@@ -5,7 +5,7 @@
 namespace Hasm::ErrorMessage
 {
 
-[[nodiscard]] std::string invalidCommand( const std::string& cmd, int lineNumber );
-[[nodiscard]] std::string invalidLoadValue( const std::string& cmd, int lineNumber );
+[[nodiscard]] std::string invalidCommand( const std::string& cmd, size_t lineNumber );
+[[nodiscard]] std::string invalidLoadValue( const std::string& cmd, size_t lineNumber );
 
 } // namespace Hasm::ErrorMessage
